requirements.txt
altgraph==0.17.3
asgiref==3.6.0
Django==4.2
ffmpeg-python==0.2.0
future==0.18.3
image==1.5.33
pefile==2023.2.7
Pillow==9.5.0
pyinstaller==5.10.1
pyinstaller-hooks-contrib==2023.2
pywin32-ctypes==0.2.0
six==1.16.0
sqlparse==0.4.4
tk==0.1.0
tkintertable==1.3.3
tzdata==2023.3
